requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
decorator==5.1.1
dnspython==2.7.0
emoji==2.14.0
executing==2.1.0
frozenlist==1.5.0
idna==3.10
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.2
magic-filter==1.0.12
matplotlib-inline==0.1.7
multidict==6.1.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.10.1
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
yarl==1.17.1
python-telegram-bot